A tibble with details on the council make-up of local authorities across the UK.

This data is taken from http://opencouncildata.co.uk/, run by Jon Lawson. Variable names have been converted to snake_case and variables have been converted to appropriate R classes. ONS local authority codes have been taken from the local_hex_map data, and from the ONS's Open Geography Portal. This data is accurate as of 2017-06-02.

council_data

Format

A tibble, with 418 rows and 20 columns.

la_code

The Local Authority code from the Office for National Statistics

name

Local Authority Name

type

The type of local authority

majority_party

The majority party on the council. The label 'No Overall Control' is used if no single party has a majority.

majority_party_id

The ID of the majority party on the council. The label 'noc' is used if no single party has a majority.

governing_coalition

The governing coalition of the council (if known)

conservative_councillors

The number of Conservative councillors

labour_councillors

The number of Labour councillors

lib_dem_councillors

The number of Liberal Democrat councillors

green_councillors

The number of Green councillors

ukip_councillors

The number of UKIP councillors

plaid_cymru_councillors

The number of Plaid Cymru councillors

snp_councillors

The number of SNP councillors

independent_councillors

The number of Independent councillors

vacant

The number of vacant seats on the council

total_councillors

Total number of vacant council seats

boundary

Notes on upcoming boundary changes

notes

Notes from Open Council Data UK

url

URL

last_update

Date of last update

Source

http://opencouncildata.co.uk/councils.php?model= http://geoportal.statistics.gov.uk/datasets/464be6191a434a91a5fa2f52c7433333_0