A tibble with the results of the 2017 general election.

A tibble with Great Britain constituencies results from the 2017 general election. This data can be linked to 2011 census information in bes_2015.

ge_2017

Format

A tibble, containing 632 row and 58 columns.

pano

Press Association Constituency ID. Numeric.

onsconst_id

ONS Constituency ID. Character.

constituency_name

Constituency Name. Factor.

country

Country. Factor.

region

Region. Factor.

constituency_type

County or Borough Constituency. Factor.

winner_17

2017 Winning party. Factor.

electorate_17

2017 Size of electorate. Numeric.

total_vote_17

2017 Total number of votes cast. Numeric.

rejected_vote_17

2017 Number of rejected ballots. Numeric.

majority_17

2017 Majority. Numeric.

majority_17

2017 Majority. Numeric.

turnout_17

2017 Turnout. Numeric.

seat_change_1517

2017 winning party gain from 2015 winning party. Factor.

con_vote_17

2017 Conservative number of votes. Numeric.

lab_vote_17

2017 Labour number of votes. Numeric.

ld_vote_17

2017 Liberal Democrat number of votes. Numeric.

ukip_vote_17

2017 United Kingdom Independence Party (UKIP) number of votes. Numeric.

green_vote_17

2017 Green Party number of votes. Numeric.

snp_vote_17

2017 Scottish National Party (SNP) number of votes. Numeric.

pc_vote_17

2017 Plaid Cymru number of votes. Numeric.

other_vote_17

2017 Other vote share. Numeric.

con_17

2017 Conservative vote share. Numeric.

lab_17

2017 Labour vote share. Numeric.

ld_17

2017 Liberal Democrat vote share. Numeric.

snp_17

2017 Scottish National Party (SNP) vote share. Numeric.

pc_17

2017 Plaid Cymru vote share. Numeric.

ukip_17

2017 United Kingdom Independence Party (UKIP) vote share. Numeric.

green_17

2017 Green Party vote share. Numeric.

other_17

2017 Combined other vote share. Numeric.

con_1517

2015-17 Conservative vote share change. Numeric.

lab_1517

2015-17 Labour vote share change. Numeric.

ld_1517

2015-17 Liberal Democrat vote share change. Numeric.

snp_1517

2015-17 Scottish National Party (SNP) vote share change. Numeric.

pc_1517

2015-17 Plaid Cymru vote share change. Numeric.

ukip_1517

2015-17 United Kingdom Independence Party (UKIP) vote share change. Numeric.

green_1517

2015-17 Green Party vote share change. Numeric.

winner_15

2015 Winning party. Factor.

con_15

2015 Conservative vote share. Numeric.

lab_15

2015 Labour vote share. Numeric.

ld_15

2015 Liberal Democrat vote share. Numeric.

snp_15

2015 Scottish National Party (SNP) vote share. Numeric.

pc_15

2015 Plaid Cymru vote share. Numeric.

ukip_15

2015 United Kingdom Independence Party (UKIP) vote share. Numeric.

green_15

2015 Green Party vote share. Numeric.

other_15

2015 Combined other vote share. Numeric.

majority_15

2015 Majority. Numeric.

turnout_15

2015 Turnout. Numeric.

con_vote_15

2015 Conservative number of votes. Numeric.

lab_vote_15

2015 Labour number of votes. Numeric.

ldvote_15

2015 Liberal Democrat number of votes. Numeric.

snpvote_15

2015 Scottish National Party (SNP) number of votes. Numeric.

pcvote_15

2015 Plaid Cymru number of votes. Numeric.

ukipvote_15

2015 United Kingdom Independence Party (UKIP) number of votes. Numeric.

green_vote_15

2015 Green Party number of votes. Numeric.

bnpvote_15

2015 British National Party (BNP) number of votes. Numeric.

total_vote_15

2015 Total number of votes cast. Numeric.

rejected_vote_15

2015 Number of rejected ballots. Numeric.

electorate_15

2015 Size of electorate. Numeric.

Source

researchbriefings.parliament.uk/ResearchBriefing/Summary/CBP-7979 http://www.britishelectionstudy.com/data-object/2015-bes-constituency-results-with-census-and-candidate-data/

Details

Variable names have been converted to snake_case and variables have been converted to appropriate R classes.