A tibble with details on percentage of votes cast for leave in the 2016 referendum on British membership of the European Union in each British constituency, excluding Northern Ireland. Numbers for Britian were compiled by the House of Commons using data from Chris Hanretty's estimates of leave vote, and supplemented with data from BBC Freedom of Information requests for more finely grained voting data. Numbers for Northern Ireland, which reported results by constituency, are taken from the BBC.

A tibble with details on percentage of votes cast for leave in the 2016 referendum on British membership of the European Union in each British constituency, excluding Northern Ireland. Numbers for Britian were compiled by the House of Commons using data from Chris Hanretty's estimates of leave vote, and supplemented with data from BBC Freedom of Information requests for more finely grained voting data. Numbers for Northern Ireland, which reported results by constituency, are taken from the BBC.

leave_votes_west

Format

A tibble, with 641 rows and 7 columns.

gss_code

ONS code for constituency

constituency

The name of the constituency.

party_2016

The party holding the constituency on 23 June 2016, the date of the referendum.

ch_leave_vote

Estimates of the leave vote produced by Chris Hanretty.

known_leave_vote

A dummy variable indicating 'yes' if the actual leave vote percentage is known, and 'no' if the estimate by CHris Hanretty is available.

known_leave_vote_perc

The percentage of votes cast for leave, if known.

difference_estimate_known

Difference between the estimate by Chris Hanretty and the actual result, if known.

Source

https://secondreading.uk/brexit/brexit-votes-by-constituency/ http://www.bbc.co.uk/news/uk-northern-ireland-36616830