All current MPs

Request data on all MPs eligible currently (i.e. on the current system date) to sit in the House of Commons. Includes information on their constituency.

current_mps(tidy = TRUE, tidy_style = "snake_case")

Arguments

tidy
If TRUE, fixes the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to TRUE.

tidy_style
The style to convert variable names to, if tidy=TRUE. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.

Value

A tibble of all MPs currently eligible to sit in the House of Commons.

Examples

if (FALSE) {
x <- current_mps()
}