A tibble with the ID, name and hex code for the official colour of a variety of political parties.

Includes all political parties with MPs and a number without MPs. It also contains a 'No Overall Control' party for local government. Hex codes are taken from Wikipedia.

party_colour

Format

A tibble, containing 47 rows and 3 columns.

party_id

The party ID, as assigned by the Members Name Information Service

party_name

The name of the political party

party_colour

A hex code for the party colour, as used by Wikipedia

Source

https://en.wikipedia.org/wiki/Wikipedia:Index_of_United_Kingdom_political_parties_meta_attributes http://data.parliament.uk/membersdataplatform/services/mnis/ReferenceData/Parties/