A tibble with the ID, name and hex code for the official colour of a variety of political parties, taken from Wikipedia. Includes all political parties with MPs and a number without MPs.

A tibble with the ID, name and hex code for the official colour of a variety of political parties, taken from Wikipedia. Includes all political parties with MPs and a number without MPs.

party_colour

Format

A tibble, containing 46 rows and 3 columns.

party_id

The party ID, as assigned by the Members Name Information Service

party_name

The name of the political party

party_colour

A hex code for the party colour, as used by Wikipedia

Source

https://en.wikipedia.org/wiki/Wikipedia:Index_of_United_Kingdom_political_parties_meta_attributes http://data.parliament.uk/membersdataplatform/services/mnis/ReferenceData/Parties/